re #16267 Modifying POJOs to match the correct data types
    Changing many POJOs data types: from long to int to match the db. Also
    Changed some String datatypes to boolean and foreign keys to use the
    class as the type.

+++ docroot/WEB-INF/src/edu/osu/cws/pass/models/AppointmentType.java
@@ -5,18 +5,20 @@
  */
 package edu.osu.cws.pass.models;
 
+
 public class AppointmentType {
-    private long id;
+    private int id;
 
     private String name;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    private void setId(long id) {
+    private void setId(int id) {
         this.id = id;
     }
+
 
     public String getName() {
         return name;
